persist: rename the `storage` module to `location`
Persist is near enough to STORAGE that this has a reasonable chance of
being confusing.
