+++ src/persist/benches/snapshot.rs
@@ -20,10 +20,10 @@ use mz_ore::metrics::MetricsRegistry;
 use mz_persist::client::{RuntimeClient, StreamReadHandle};
 use mz_persist::error::{Error, ErrorLog};
 use mz_persist::file::FileBlob;
+use mz_persist::location::{Atomicity, Blob, BlobRead, LockInfo};
 use mz_persist::mem::MemRegistry;
 use mz_persist::runtime::{self, RuntimeConfig};
 use mz_persist::s3::{S3Blob, S3BlobConfig};
-use mz_persist::storage::{Atomicity, Blob, BlobRead, LockInfo};
 use mz_persist::workload::{self, DataGenerator};
 use mz_persist_types::Codec;
 use uuid::Uuid;
