+++ src/persist/src/lib.rs
@@ -25,13 +25,13 @@ pub mod error;
 pub mod file;
 pub mod gen;
 pub mod indexed;
+pub mod location;
 pub mod mem;
 pub mod operators;
 pub mod pfuture;
 pub mod runtime;
 pub mod s3;
 pub mod sqlite;
-pub mod storage;
 pub mod unreliable;
 pub mod workload;
