+++ src/persist/src/nemesis/direct.rs
@@ -109,6 +109,7 @@ use crate::client::{
     DecodedSnapshot, MultiWriteHandle, RuntimeClient, StreamReadHandle, StreamWriteHandle,
 };
 use crate::error::Error;
+use crate::location::SeqNo;
 use crate::nemesis::progress::{DataflowProgress, DataflowProgressHandle};
 use crate::nemesis::{
     AllowCompactionReq, FutureRes, FutureStep, Input, ReadOutputEvent, ReadOutputReq,
@@ -118,7 +119,6 @@ use crate::nemesis::{
 };
 use crate::operators::source::PersistedSource;
 use crate::pfuture::PFuture;
-use crate::storage::SeqNo;
 use crate::unreliable::UnreliableHandle;
 
 /// A handle to the "ingestd" (input) side of a persisted collection.
@@ -569,11 +569,11 @@ mod tests {
 
     use crate::error::ErrorLog;
     use crate::file::{FileBlob, FileLog};
+    use crate::location::Blob;
     use crate::mem::MemRegistry;
     use crate::nemesis::generator::GeneratorConfig;
     use crate::runtime::RuntimeConfig;
     use crate::s3::{S3Blob, S3BlobConfig};
-    use crate::storage::Blob;
     use crate::unreliable::{UnreliableBlob, UnreliableLog};
     use crate::{nemesis, runtime};
